Use `long` for distance calculations to prevent overflow.
Updated `SantaTravelCalculator` to use `long` instead of `int` for distance calculations, ensuring support for larger inputs. Adjusted unit tests to validate the behavior with higher values, removing outdated tests for overflow scenarios.

+++ exercise/C#/day19/Travel/SantaTravelCalculator.cs
@@ -2,13 +2,13 @@
 {
 public static class SantaTravelCalculator
 {
-    public static int CalculateTotalDistanceRecursively(int numberOfReindeers)
+    public static long CalculateTotalDistanceRecursively(int numberOfReindeers)
     {
         if (numberOfReindeers == 1) return 1;
         checked
         {
-            return (int)Math
-                .Pow(2, numberOfReindeers) - 1;
+            return (long)(Math
+                .Pow(2, numberOfReindeers) - 1);
         }
     }
 }
